move random seed to setup
Moved randomSeed to setup routine as this gave better randomness. On my board, the analog pins don't appear to be very good so doing the randomSeed once gives better results than seeding it each time a new question is calculated. 

Another alternative would be to do randomSeed at the start of each game.